Working on requests

+++ backend/src/main/java/com/nutriometer/repository/RecipeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecipeRepository extends CrudRepository<Recipe, Integer> {
+public interface RecipeRepository extends CrudRepository<Recipe, Long> {
     Recipe findByName(String recipeName);
 
     Iterable<Recipe> findAll();
